+++ initialize.sh
@@ -19,8 +19,8 @@ echo "Attempting to build"
 gradle wrapper --gradle-version 8.6
 ./gradlew build
 
-echo "Checking if Bip32Ed25519-Android*.aar and Bip32Ed25519-JVM.jar exist in build/ directory..."
-for pattern in 'build/Bip32Ed25519-Android*.aar' 'build/Bip32Ed25519-JVM.jar'; do
+echo "Checking if XHDWalletAPI-Android*.aar and XHDWalletAPI-JVM.jar exist in build/ directory..."
+for pattern in 'build/XHDWalletAPI-Android*.aar' 'build/XHDWalletAPI-JVM.jar'; do
     files=( $pattern )
     if [ -e "${files[0]}" ]; then
         echo "Build files matching expected $pattern exist."
