+++ jvmModule/src/main/kotlin/com/algorandfoundation/xhdwalletapi/XHDWalletAPIJVM.kt
@@ -15,11 +15,12 @@
  * limitations under the License.
  */
 
-package com.algorandfoundation.bip32ed25519
+package com.algorandfoundation.xhdwalletapi
 
-import com.goterl.lazysodium.LazySodiumAndroid
-import com.goterl.lazysodium.SodiumAndroid
+import com.goterl.lazysodium.LazySodiumJava
+import com.goterl.lazysodium.SodiumJava
+import com.goterl.lazysodium.utils.LibraryLoader
 
-class Bip32Ed25519Android(private var seed: ByteArray) : Bip32Ed25519Base(seed) {
-    override val lazySodium = LazySodiumAndroid(SodiumAndroid())
+open class XHDWalletAPIJVM(private var seed: ByteArray) : XHDWalletAPIBase(seed) {
+    override val lazySodium = LazySodiumJava(SodiumJava(LibraryLoader.Mode.PREFER_BUNDLED))
 }
